Apply reunião 25/06 changes: report layout, horário lock, packs self-view
Implements items 1–5 of the meeting plan (plano-reuniao2506):

- Desempenho pct/totalPct int→double, rounded to 2 decimals; period total
  is now the mean of daily % (days with meta>0) instead of production-weighted,
  matching the per-day "média das %" rule.
- Report layout (admin Detalhe do dia + portal): Meta | Produção | % columns
  side by side per operation, drop the summed-meta total (keep produced pieces
  + average %), 2-decimal percentages everywhere.
- Horário lock / two user types: ADMIN edits the start time freely; OPERADOR
  gets a read-only, server-stamped horário (AlocacaoService stamps now on create
  and freezes it on update). Stamp uses ezcostura.timezone (America/Sao_Paulo by
  default) since the prod JVM runs UTC.
- Default start time from the jornada: 1st allocation of the day starts at the
  shift start, 2nd+ at end of lunch; the facilitador (real-time) context uses now.
- Operário portal packs self-view: GET /api/me/packs returns the worker's own
  packs (horário, quantidade, tamanho); portal rows expand to list them.

+++ backend/src/main/java/com/ezcostura/desempenho/dto/DesempenhoDiaDto.java
@@ -16,5 +16,5 @@ public record DesempenhoDiaDto(
     List<DesempenhoItemDto> itens,
     int totalMeta,
     int totalProduzido,
-    int totalPct
+    double totalPct
 ) {}
